requirements.txt
fastapi==0.120.4
uvicorn==0.38.0
faiss-cpu==1.12.0
sentence-transformers==5.1.2
ollama==0.6.0
torch==2.9.0
transformers==4.57.1
scikit-learn==1.7.2
